refactor: moved exercises methods to new files
